+++ src/components/HomePage.tsx
@@ -1,0 +1,8 @@
+import React from 'react';
+
+
+export const HomePage = () => {
+  return (
+    <h2>Home Page perhaps</h2>
+  );
+};
